Extract middleware test from utils.spec.ts
