+++ specs/utils.spec.ts
@@ -3,81 +3,12 @@ import { testHandler } from './testHandler'
 import {
   prismy,
   res,
-  Selector,
-  PrismyPureMiddleware,
-  middleware,
   redirect,
   setBody,
   setStatusCode,
   updateHeaders,
-  setHeaders,
-  AsyncSelector
+  setHeaders
 } from '../src'
-
-describe('middleware', () => {
-  it('creates Middleware via selectors and middleware handler', async () => {
-    const rawUrlSelector: Selector<string> = context => context.req.url!
-    const errorMiddleware: PrismyPureMiddleware = middleware(
-      [rawUrlSelector],
-      next => async url => {
-        try {
-          return await next()
-        } catch (error) {
-          return res(`${url} : ${error.message}`, 500)
-        }
-      }
-    )
-    const handler = prismy(
-      [],
-      () => {
-        throw new Error('Hey!')
-      },
-      [errorMiddleware]
-    )
-
-    await testHandler(handler, async url => {
-      const response = await got(url, {
-        throwHttpErrors: false
-      })
-      expect(response).toMatchObject({
-        statusCode: 500,
-        body: '/ : Hey!'
-      })
-    })
-  })
-
-  it('accepts async selectors', async () => {
-    const asyncRawUrlSelector: AsyncSelector<string> = async context =>
-      context.req.url!
-    const errorMiddleware = middleware(
-      [asyncRawUrlSelector],
-      next => async url => {
-        try {
-          return await next()
-        } catch (error) {
-          return res(`${url} : ${error.message}`, 500)
-        }
-      }
-    )
-    const handler = prismy(
-      [],
-      () => {
-        throw new Error('Hey!')
-      },
-      [errorMiddleware]
-    )
-
-    await testHandler(handler, async url => {
-      const response = await got(url, {
-        throwHttpErrors: false
-      })
-      expect(response).toMatchObject({
-        statusCode: 500,
-        body: '/ : Hey!'
-      })
-    })
-  })
-})
 
 describe('redirect', () => {
   it('redirects', async () => {
